src/ol_superset/pyproject.toml
[project]
name = "ol-superset"
version = "0.1.0"
description = "CLI for managing Apache Superset assets across environments"
requires-python = ">=3.13"
dependencies = [
    "cyclopts>=3.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "superset-sup",
]

[project.scripts]
ol-superset = "ol_superset.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["ol_superset"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv.sources]
superset-sup = { git = "https://github.com/mitodl/superset-sup", rev = "self_hosted_superset_oidc" }

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "S", "B", "UP"]
ignore = ["S603", "S607"]  # Allow subprocess calls to sup CLI

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert is expected in pytest tests
